move evm_circuit::witness as crate::witness; split it into smaller mods
