+++ zkevm-circuits/src/evm_circuit.rs
@@ -5,13 +5,13 @@ use halo2_proofs::{circuit::Layouter, plonk::*};
 
 mod execution;
 pub mod param;
-mod step;
+pub(crate) mod step;
 pub(crate) mod util;
 
 pub mod table;
-pub mod witness;
 
 use crate::table::LookupTable;
+pub use crate::witness;
 use eth_types::Field;
 use execution::ExecutionConfig;
 use itertools::Itertools;
