+++ zkevm-circuits/src/witness.rs
@@ -1,0 +1,14 @@
+#![allow(missing_docs)]
+
+mod block;
+pub use block::{block_convert, Block, BlockContext};
+mod bytecode;
+pub use bytecode::Bytecode;
+mod call;
+pub use call::Call;
+mod rw;
+pub use rw::{Rw, RwMap, RwRow};
+mod step;
+pub use step::ExecStep;
+mod tx;
+pub use tx::Transaction;
