+++ zkevm-circuits/src/witness/rw.rs
@@ -1,376 +1,13 @@
-#![allow(missing_docs)]
-use crate::{
-    evm_circuit::{
-        param::{N_BYTES_WORD, STACK_CAPACITY},
-        step::ExecutionState,
-        util::RandomLinearCombination,
-    },
-    table::{
-        AccountFieldTag, BlockContextFieldTag, BytecodeFieldTag, CallContextFieldTag, RwTableTag,
-        TxContextFieldTag, TxLogFieldTag, TxReceiptFieldTag,
-    },
-};
+use std::collections::HashMap;
 
-use bus_mapping::{
-    circuit_input_builder::{self, CopyEvent},
-    error::{ExecError, OogError},
-    operation::{self, AccountField, CallContextField, TxLogField, TxReceiptField},
-};
-
-use eth_types::{evm_types::OpcodeId, ToWord};
-use eth_types::{Address, Field, ToLittleEndian, ToScalar, Word};
-use eth_types::{ToAddress, U256};
-use halo2_proofs::arithmetic::FieldExt;
-use halo2_proofs::pairing::bn256::Fr;
+use bus_mapping::operation::{self, AccountField, CallContextField, TxLogField, TxReceiptField};
+use eth_types::{Address, Field, ToAddress, ToLittleEndian, ToScalar, Word, U256};
 use itertools::Itertools;
-use sha3::{Digest, Keccak256};
-use std::{collections::HashMap, iter};
 
-#[derive(Debug, Default, Clone)]
-pub struct Block<F> {
-    /// The randomness for random linear combination
-    pub randomness: F,
-    /// Transactions in the block
-    pub txs: Vec<Transaction>,
-    /// Read write events in the RwTable
-    pub rws: RwMap,
-    /// Bytecode used in the block
-    pub bytecodes: HashMap<Word, Bytecode>,
-    /// The block context
-    pub context: BlockContext,
-    /// Copy events for the EVM circuit's copy table.
-    pub copy_events: Vec<CopyEvent>,
-    /// Length to rw table rows in state circuit
-    pub state_circuit_pad_to: usize,
-    /// Inputs to the SHA3 opcode
-    pub sha3_inputs: Vec<Vec<u8>>,
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct BlockContext {
-    /// The address of the miner for the block
-    pub coinbase: Address,
-    /// The gas limit of the block
-    pub gas_limit: u64,
-    /// The number of the block
-    pub number: Word,
-    /// The timestamp of the block
-    pub timestamp: Word,
-    /// The difficulty of the blcok
-    pub difficulty: Word,
-    /// The base fee, the minimum amount of gas fee for a transaction
-    pub base_fee: Word,
-    /// The hash of previous blocks
-    pub history_hashes: Vec<Word>,
-    /// The chain id
-    pub chain_id: Word,
-}
-
-impl BlockContext {
-    pub fn table_assignments<F: Field>(&self, randomness: F) -> Vec<[F; 3]> {
-        [
-            vec![
-                [
-                    F::from(BlockContextFieldTag::Coinbase as u64),
-                    F::zero(),
-                    self.coinbase.to_scalar().unwrap(),
-                ],
-                [
-                    F::from(BlockContextFieldTag::Timestamp as u64),
-                    F::zero(),
-                    self.timestamp.to_scalar().unwrap(),
-                ],
-                [
-                    F::from(BlockContextFieldTag::Number as u64),
-                    F::zero(),
-                    self.number.to_scalar().unwrap(),
-                ],
-                [
-                    F::from(BlockContextFieldTag::Difficulty as u64),
-                    F::zero(),
-                    RandomLinearCombination::random_linear_combine(
-                        self.difficulty.to_le_bytes(),
-                        randomness,
-                    ),
-                ],
-                [
-                    F::from(BlockContextFieldTag::GasLimit as u64),
-                    F::zero(),
-                    F::from(self.gas_limit),
-                ],
-                [
-                    F::from(BlockContextFieldTag::BaseFee as u64),
-                    F::zero(),
-                    RandomLinearCombination::random_linear_combine(
-                        self.base_fee.to_le_bytes(),
-                        randomness,
-                    ),
-                ],
-                [
-                    F::from(BlockContextFieldTag::ChainId as u64),
-                    F::zero(),
-                    RandomLinearCombination::random_linear_combine(
-                        self.chain_id.to_le_bytes(),
-                        randomness,
-                    ),
-                ],
-            ],
-            self.history_hashes
-                .iter()
-                .rev()
-                .enumerate()
-                .map(|(idx, hash)| {
-                    [
-                        F::from(BlockContextFieldTag::BlockHash as u64),
-                        (self.number - idx - 1).to_scalar().unwrap(),
-                        RandomLinearCombination::random_linear_combine(
-                            hash.to_le_bytes(),
-                            randomness,
-                        ),
-                    ]
-                })
-                .collect(),
-        ]
-        .concat()
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct Transaction {
-    /// The transaction identifier in the block
-    pub id: usize,
-    /// The sender account nonce of the transaction
-    pub nonce: u64,
-    /// The gas limit of the transaction
-    pub gas: u64,
-    /// The gas price
-    pub gas_price: Word,
-    /// The caller address
-    pub caller_address: Address,
-    /// The callee address
-    pub callee_address: Address,
-    /// Whether it's a create transaction
-    pub is_create: bool,
-    /// The ether amount of the transaction
-    pub value: Word,
-    /// The call data
-    pub call_data: Vec<u8>,
-    /// The call data length
-    pub call_data_length: usize,
-    /// The gas cost for transaction call data
-    pub call_data_gas_cost: u64,
-    /// The calls made in the transaction
-    pub calls: Vec<Call>,
-    /// The steps executioned in the transaction
-    pub steps: Vec<ExecStep>,
-}
-
-impl Transaction {
-    pub fn table_assignments<F: Field>(&self, randomness: F) -> Vec<[F; 4]> {
-        [
-            vec![
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::Nonce as u64),
-                    F::zero(),
-                    F::from(self.nonce),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::Gas as u64),
-                    F::zero(),
-                    F::from(self.gas),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::GasPrice as u64),
-                    F::zero(),
-                    RandomLinearCombination::random_linear_combine(
-                        self.gas_price.to_le_bytes(),
-                        randomness,
-                    ),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::CallerAddress as u64),
-                    F::zero(),
-                    self.caller_address.to_scalar().unwrap(),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::CalleeAddress as u64),
-                    F::zero(),
-                    self.callee_address.to_scalar().unwrap(),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::IsCreate as u64),
-                    F::zero(),
-                    F::from(self.is_create as u64),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::Value as u64),
-                    F::zero(),
-                    RandomLinearCombination::random_linear_combine(
-                        self.value.to_le_bytes(),
-                        randomness,
-                    ),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::CallDataLength as u64),
-                    F::zero(),
-                    F::from(self.call_data_length as u64),
-                ],
-                [
-                    F::from(self.id as u64),
-                    F::from(TxContextFieldTag::CallDataGasCost as u64),
-                    F::zero(),
-                    F::from(self.call_data_gas_cost),
-                ],
-            ],
-            self.call_data
-                .iter()
-                .enumerate()
-                .map(|(idx, byte)| {
-                    [
-                        F::from(self.id as u64),
-                        F::from(TxContextFieldTag::CallData as u64),
-                        F::from(idx as u64),
-                        F::from(*byte as u64),
-                    ]
-                })
-                .collect(),
-        ]
-        .concat()
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct Call {
-    /// The unique identifier of call in the whole proof, using the
-    /// `rw_counter` at the call step.
-    pub id: usize,
-    /// Indicate if the call is the root call
-    pub is_root: bool,
-    /// Indicate if the call is a create call
-    pub is_create: bool,
-    /// The identifier of current executed bytecode
-    pub code_hash: Word,
-    /// The `rw_counter` at the end of reversion of a call if it has
-    /// `is_persistent == false`
-    pub rw_counter_end_of_reversion: usize,
-    /// The call index of caller
-    pub caller_id: usize,
-    /// The depth in the call stack
-    pub depth: usize,
-    /// The caller address
-    pub caller_address: Address,
-    /// The callee address
-    pub callee_address: Address,
-    /// The call data offset in the memory
-    pub call_data_offset: u64,
-    /// The length of call data
-    pub call_data_length: u64,
-    /// The return data offset in the memory
-    pub return_data_offset: u64,
-    /// The length of return data
-    pub return_data_length: u64,
-    /// The ether amount of the transaction
-    pub value: Word,
-    /// Indicate if this call halts successfully.
-    pub is_success: bool,
-    /// Indicate if this call and all its caller have `is_success == true`
-    pub is_persistent: bool,
-    /// Indicate if it's a static call
-    pub is_static: bool,
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct ExecStep {
-    /// The index in the Transaction calls
-    pub call_index: usize,
-    /// The indices in the RW trace incurred in this step
-    pub rw_indices: Vec<(RwTableTag, usize)>,
-    /// The execution state for the step
-    pub execution_state: ExecutionState,
-    /// The Read/Write counter before the step
-    pub rw_counter: usize,
-    /// The program counter
-    pub program_counter: u64,
-    /// The stack pointer
-    pub stack_pointer: usize,
-    /// The amount of gas left
-    pub gas_left: u64,
-    /// The gas cost in this step
-    pub gas_cost: u64,
-    /// The memory size in bytes
-    pub memory_size: u64,
-    /// The counter for reversible writes
-    pub reversible_write_counter: usize,
-    /// The counter for log index within tx
-    pub log_id: usize,
-    /// The opcode corresponds to the step
-    pub opcode: Option<OpcodeId>,
-}
-
-impl ExecStep {
-    pub fn memory_word_size(&self) -> u64 {
-        // EVM always pads the memory size to word size
-        // https://github.com/ethereum/go-ethereum/blob/master/core/vm/interpreter.go#L212-L216
-        // Thus, the memory size must be a multiple of 32 bytes.
-        assert_eq!(self.memory_size % N_BYTES_WORD as u64, 0);
-        self.memory_size / N_BYTES_WORD as u64
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct Bytecode {
-    pub hash: Word,
-    pub bytes: Vec<u8>,
-}
-
-impl Bytecode {
-    pub fn new(bytes: Vec<u8>) -> Self {
-        let hash = Word::from_big_endian(Keccak256::digest(&bytes).as_slice());
-        Self { hash, bytes }
-    }
-
-    pub fn table_assignments<F: FieldExt>(&self, randomness: F) -> Vec<[F; 5]> {
-        let n = 1 + self.bytes.len();
-        let mut rows = Vec::with_capacity(n);
-        let hash =
-            RandomLinearCombination::random_linear_combine(self.hash.to_le_bytes(), randomness);
-
-        rows.push([
-            hash,
-            F::from(BytecodeFieldTag::Length as u64),
-            F::zero(),
-            F::zero(),
-            F::from(self.bytes.len() as u64),
-        ]);
-
-        let mut push_data_left = 0;
-        for (idx, byte) in self.bytes.iter().enumerate() {
-            let mut is_code = true;
-            if push_data_left > 0 {
-                is_code = false;
-                push_data_left -= 1;
-            } else if (OpcodeId::PUSH1.as_u8()..=OpcodeId::PUSH32.as_u8()).contains(byte) {
-                push_data_left = *byte as usize - (OpcodeId::PUSH1.as_u8() - 1) as usize;
-            }
-            rows.push([
-                hash,
-                F::from(BytecodeFieldTag::Byte as u64),
-                F::from(idx as u64),
-                F::from(is_code as u64),
-                F::from(*byte as u64),
-            ])
-        }
-        rows
-    }
-}
+use crate::{
+    evm_circuit::util::RandomLinearCombination,
+    table::{AccountFieldTag, CallContextFieldTag, RwTableTag, TxLogFieldTag, TxReceiptFieldTag},
+};
 
 #[derive(Debug, Default, Clone)]
 pub struct RwMap(pub HashMap<RwTableTag, Vec<Rw>>);
@@ -931,21 +568,6 @@ impl Rw {
     }
 }
 
-impl From<&circuit_input_builder::Block> for BlockContext {
-    fn from(block: &circuit_input_builder::Block) -> Self {
-        Self {
-            coinbase: block.coinbase,
-            gas_limit: block.gas_limit,
-            number: block.number,
-            timestamp: block.timestamp,
-            difficulty: block.difficulty,
-            base_fee: block.base_fee,
-            history_hashes: block.history_hashes.clone(),
-            chain_id: block.chain_id,
-        }
-    }
-}
-
 impl From<&operation::OperationContainer> for RwMap {
     fn from(container: &operation::OperationContainer) -> Self {
         let mut rws = HashMap::default();
@@ -1164,293 +786,5 @@ impl From<&operation::OperationContainer> for RwMap {
         );
 
         Self(rws)
-    }
-}
-
-impl From<&ExecError> for ExecutionState {
-    fn from(error: &ExecError) -> Self {
-        match error {
-            ExecError::InvalidOpcode => ExecutionState::ErrorInvalidOpcode,
-            ExecError::StackOverflow => ExecutionState::ErrorStackOverflow,
-            ExecError::StackUnderflow => ExecutionState::ErrorStackUnderflow,
-            ExecError::WriteProtection => ExecutionState::ErrorWriteProtection,
-            ExecError::Depth => ExecutionState::ErrorDepth,
-            ExecError::InsufficientBalance => ExecutionState::ErrorInsufficientBalance,
-            ExecError::ContractAddressCollision => ExecutionState::ErrorContractAddressCollision,
-            ExecError::InvalidCreationCode => ExecutionState::ErrorInvalidCreationCode,
-            ExecError::InvalidJump => ExecutionState::ErrorInvalidJump,
-            ExecError::ReturnDataOutOfBounds => ExecutionState::ErrorReturnDataOutOfBound,
-            ExecError::CodeStoreOutOfGas => ExecutionState::ErrorOutOfGasCodeStore,
-            ExecError::MaxCodeSizeExceeded => ExecutionState::ErrorMaxCodeSizeExceeded,
-            ExecError::OutOfGas(oog_error) => match oog_error {
-                OogError::Constant => ExecutionState::ErrorOutOfGasConstant,
-                OogError::StaticMemoryExpansion => {
-                    ExecutionState::ErrorOutOfGasStaticMemoryExpansion
-                }
-                OogError::DynamicMemoryExpansion => {
-                    ExecutionState::ErrorOutOfGasDynamicMemoryExpansion
-                }
-                OogError::MemoryCopy => ExecutionState::ErrorOutOfGasMemoryCopy,
-                OogError::AccountAccess => ExecutionState::ErrorOutOfGasAccountAccess,
-                OogError::CodeStore => ExecutionState::ErrorOutOfGasCodeStore,
-                OogError::Log => ExecutionState::ErrorOutOfGasLOG,
-                OogError::Exp => ExecutionState::ErrorOutOfGasEXP,
-                OogError::Sha3 => ExecutionState::ErrorOutOfGasSHA3,
-                OogError::ExtCodeCopy => ExecutionState::ErrorOutOfGasEXTCODECOPY,
-                OogError::Sload => ExecutionState::ErrorOutOfGasSLOAD,
-                OogError::Sstore => ExecutionState::ErrorOutOfGasSSTORE,
-                OogError::Call => ExecutionState::ErrorOutOfGasCALL,
-                OogError::CallCode => ExecutionState::ErrorOutOfGasCALLCODE,
-                OogError::DelegateCall => ExecutionState::ErrorOutOfGasDELEGATECALL,
-                OogError::Create2 => ExecutionState::ErrorOutOfGasCREATE2,
-                OogError::StaticCall => ExecutionState::ErrorOutOfGasSTATICCALL,
-                OogError::SelfDestruct => ExecutionState::ErrorOutOfGasSELFDESTRUCT,
-            },
-        }
-    }
-}
-
-impl From<&circuit_input_builder::ExecStep> for ExecutionState {
-    fn from(step: &circuit_input_builder::ExecStep) -> Self {
-        if let Some(error) = step.error.as_ref() {
-            return error.into();
-        }
-        match step.exec_state {
-            circuit_input_builder::ExecState::Op(op) => {
-                if op.is_dup() {
-                    return ExecutionState::DUP;
-                }
-                if op.is_push() {
-                    return ExecutionState::PUSH;
-                }
-                if op.is_swap() {
-                    return ExecutionState::SWAP;
-                }
-                if op.is_log() {
-                    return ExecutionState::LOG;
-                }
-
-                macro_rules! dummy {
-                    ($name:expr) => {{
-                        log::warn!("{:?} is implemented with DummyGadget", $name);
-                        $name
-                    }};
-                }
-
-                match op {
-                    OpcodeId::ADD | OpcodeId::SUB => ExecutionState::ADD_SUB,
-                    OpcodeId::ADDMOD => ExecutionState::ADDMOD,
-                    OpcodeId::ADDRESS => ExecutionState::ADDRESS,
-                    OpcodeId::MUL | OpcodeId::DIV | OpcodeId::MOD => ExecutionState::MUL_DIV_MOD,
-                    OpcodeId::MULMOD => ExecutionState::MULMOD,
-                    OpcodeId::SDIV | OpcodeId::SMOD => ExecutionState::SDIV_SMOD,
-                    OpcodeId::EQ | OpcodeId::LT | OpcodeId::GT => ExecutionState::CMP,
-                    OpcodeId::SLT | OpcodeId::SGT => ExecutionState::SCMP,
-                    OpcodeId::SIGNEXTEND => ExecutionState::SIGNEXTEND,
-                    OpcodeId::STOP => ExecutionState::STOP,
-                    OpcodeId::AND => ExecutionState::BITWISE,
-                    OpcodeId::XOR => ExecutionState::BITWISE,
-                    OpcodeId::OR => ExecutionState::BITWISE,
-                    OpcodeId::NOT => ExecutionState::NOT,
-                    OpcodeId::POP => ExecutionState::POP,
-                    OpcodeId::PUSH32 => ExecutionState::PUSH,
-                    OpcodeId::BYTE => ExecutionState::BYTE,
-                    OpcodeId::MLOAD => ExecutionState::MEMORY,
-                    OpcodeId::MSTORE => ExecutionState::MEMORY,
-                    OpcodeId::MSTORE8 => ExecutionState::MEMORY,
-                    OpcodeId::JUMPDEST => ExecutionState::JUMPDEST,
-                    OpcodeId::JUMP => ExecutionState::JUMP,
-                    OpcodeId::JUMPI => ExecutionState::JUMPI,
-                    OpcodeId::GASPRICE => ExecutionState::GASPRICE,
-                    OpcodeId::PC => ExecutionState::PC,
-                    OpcodeId::MSIZE => ExecutionState::MSIZE,
-                    OpcodeId::CALLER => ExecutionState::CALLER,
-                    OpcodeId::CALLVALUE => ExecutionState::CALLVALUE,
-                    OpcodeId::EXTCODEHASH => ExecutionState::EXTCODEHASH,
-                    OpcodeId::BLOCKHASH => ExecutionState::BLOCKHASH,
-                    OpcodeId::TIMESTAMP | OpcodeId::NUMBER | OpcodeId::GASLIMIT => {
-                        ExecutionState::BLOCKCTXU64
-                    }
-                    OpcodeId::COINBASE => ExecutionState::BLOCKCTXU160,
-                    OpcodeId::DIFFICULTY | OpcodeId::BASEFEE => ExecutionState::BLOCKCTXU256,
-                    OpcodeId::GAS => ExecutionState::GAS,
-                    OpcodeId::SELFBALANCE => ExecutionState::SELFBALANCE,
-                    OpcodeId::SHA3 => ExecutionState::SHA3,
-                    OpcodeId::SHL | OpcodeId::SHR => ExecutionState::SHL_SHR,
-                    OpcodeId::SLOAD => ExecutionState::SLOAD,
-                    OpcodeId::SSTORE => ExecutionState::SSTORE,
-                    OpcodeId::CALLDATASIZE => ExecutionState::CALLDATASIZE,
-                    OpcodeId::CALLDATACOPY => ExecutionState::CALLDATACOPY,
-                    OpcodeId::CHAINID => ExecutionState::CHAINID,
-                    OpcodeId::ISZERO => ExecutionState::ISZERO,
-                    OpcodeId::CALL => ExecutionState::CALL,
-                    OpcodeId::ORIGIN => ExecutionState::ORIGIN,
-                    OpcodeId::CODECOPY => ExecutionState::CODECOPY,
-                    OpcodeId::CALLDATALOAD => ExecutionState::CALLDATALOAD,
-                    OpcodeId::CODESIZE => ExecutionState::CODESIZE,
-                    OpcodeId::RETURN | OpcodeId::REVERT => ExecutionState::RETURN,
-                    // dummy ops
-                    OpcodeId::BALANCE => dummy!(ExecutionState::BALANCE),
-                    OpcodeId::EXP => dummy!(ExecutionState::EXP),
-                    OpcodeId::SAR => dummy!(ExecutionState::SAR),
-                    OpcodeId::EXTCODESIZE => dummy!(ExecutionState::EXTCODESIZE),
-                    OpcodeId::EXTCODECOPY => dummy!(ExecutionState::EXTCODECOPY),
-                    OpcodeId::RETURNDATASIZE => dummy!(ExecutionState::RETURNDATASIZE),
-                    OpcodeId::RETURNDATACOPY => dummy!(ExecutionState::RETURNDATACOPY),
-                    OpcodeId::CREATE => dummy!(ExecutionState::CREATE),
-                    OpcodeId::CALLCODE => dummy!(ExecutionState::CALLCODE),
-                    OpcodeId::DELEGATECALL => dummy!(ExecutionState::DELEGATECALL),
-                    OpcodeId::CREATE2 => dummy!(ExecutionState::CREATE2),
-                    OpcodeId::STATICCALL => dummy!(ExecutionState::STATICCALL),
-                    OpcodeId::SELFDESTRUCT => dummy!(ExecutionState::SELFDESTRUCT),
-                    _ => unimplemented!("unimplemented opcode {:?}", op),
-                }
-            }
-            circuit_input_builder::ExecState::BeginTx => ExecutionState::BeginTx,
-            circuit_input_builder::ExecState::EndTx => ExecutionState::EndTx,
-        }
-    }
-}
-
-impl From<&eth_types::bytecode::Bytecode> for Bytecode {
-    fn from(b: &eth_types::bytecode::Bytecode) -> Self {
-        Bytecode::new(b.to_vec())
-    }
-}
-
-fn step_convert(step: &circuit_input_builder::ExecStep) -> ExecStep {
-    ExecStep {
-        call_index: step.call_index,
-        rw_indices: step
-            .bus_mapping_instance
-            .iter()
-            .map(|x| {
-                let tag = match x.target() {
-                    operation::Target::Memory => RwTableTag::Memory,
-                    operation::Target::Stack => RwTableTag::Stack,
-                    operation::Target::Storage => RwTableTag::AccountStorage,
-                    operation::Target::TxAccessListAccount => RwTableTag::TxAccessListAccount,
-                    operation::Target::TxAccessListAccountStorage => {
-                        RwTableTag::TxAccessListAccountStorage
-                    }
-                    operation::Target::TxRefund => RwTableTag::TxRefund,
-                    operation::Target::Account => RwTableTag::Account,
-                    operation::Target::AccountDestructed => RwTableTag::AccountDestructed,
-                    operation::Target::CallContext => RwTableTag::CallContext,
-                    operation::Target::TxReceipt => RwTableTag::TxReceipt,
-                    operation::Target::TxLog => RwTableTag::TxLog,
-                };
-                (tag, x.as_usize())
-            })
-            .collect(),
-        execution_state: ExecutionState::from(step),
-        rw_counter: usize::from(step.rwc),
-        program_counter: usize::from(step.pc) as u64,
-        stack_pointer: STACK_CAPACITY - step.stack_size,
-        gas_left: step.gas_left.0,
-        gas_cost: step.gas_cost.as_u64(),
-        opcode: match step.exec_state {
-            circuit_input_builder::ExecState::Op(op) => Some(op),
-            _ => None,
-        },
-        memory_size: step.memory_size as u64,
-        reversible_write_counter: step.reversible_write_counter,
-        log_id: step.log_id,
-    }
-}
-
-fn tx_convert(tx: &circuit_input_builder::Transaction, id: usize, is_last_tx: bool) -> Transaction {
-    Transaction {
-        id,
-        nonce: tx.nonce,
-        gas: tx.gas,
-        gas_price: tx.gas_price,
-        caller_address: tx.from,
-        callee_address: tx.to,
-        is_create: tx.is_create(),
-        value: tx.value,
-        call_data: tx.input.clone(),
-        call_data_length: tx.input.len(),
-        call_data_gas_cost: tx
-            .input
-            .iter()
-            .fold(0, |acc, byte| acc + if *byte == 0 { 4 } else { 16 }),
-        calls: tx
-            .calls()
-            .iter()
-            .map(|call| Call {
-                id: call.call_id,
-                is_root: call.is_root,
-                is_create: call.is_create(),
-                code_hash: call.code_hash.to_word(),
-                rw_counter_end_of_reversion: call.rw_counter_end_of_reversion,
-                caller_id: call.caller_id,
-                depth: call.depth,
-                caller_address: call.caller_address,
-                callee_address: call.address,
-                call_data_offset: call.call_data_offset,
-                call_data_length: call.call_data_length,
-                return_data_offset: call.return_data_offset,
-                return_data_length: call.return_data_length,
-                value: call.value,
-                is_success: call.is_success,
-                is_persistent: call.is_persistent,
-                is_static: call.is_static,
-            })
-            .collect(),
-        steps: tx
-            .steps()
-            .iter()
-            .map(step_convert)
-            .chain(
-                (if is_last_tx {
-                    Some(iter::once(ExecStep {
-                        // if it is the first tx,  less 1 rw lookup, refer to end_tx gadget
-                        rw_counter: tx.steps().last().unwrap().rwc.0 + 9 - (id == 1) as usize,
-                        execution_state: ExecutionState::EndBlock,
-                        ..Default::default()
-                    }))
-                } else {
-                    None
-                })
-                .into_iter()
-                .flatten(),
-            )
-            .collect(),
-    }
-}
-
-pub fn block_convert(
-    block: &circuit_input_builder::Block,
-    code_db: &bus_mapping::state_db::CodeDB,
-) -> Block<Fr> {
-    Block {
-        randomness: Fr::from(0xcafeu64),
-        context: block.into(),
-        rws: RwMap::from(&block.container),
-        txs: block
-            .txs()
-            .iter()
-            .enumerate()
-            .map(|(idx, tx)| tx_convert(tx, idx + 1, idx + 1 == block.txs().len()))
-            .collect(),
-        bytecodes: block
-            .txs()
-            .iter()
-            .flat_map(|tx| {
-                tx.calls()
-                    .iter()
-                    .map(|call| call.code_hash)
-                    .unique()
-                    .into_iter()
-                    .map(|code_hash| {
-                        let bytecode = Bytecode::new(code_db.0.get(&code_hash).unwrap().to_vec());
-                        (bytecode.hash, bytecode)
-                    })
-            })
-            .collect(),
-        copy_events: block.copy_events.clone(),
-        sha3_inputs: block.sha3_inputs.clone(),
-        ..Default::default()
     }
 }
